# [target allocator] Use ScratchBuilder for building targets