--- a/cmd/otel-allocator/internal/target/discovery.go
+++ b/cmd/otel-allocator/internal/target/discovery.go
@@ -24,6 +24,8 @@ import (
 	allocatorWatcher "github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/internal/watcher"
 )
 
+const labelBuilderPreallocSize = 100
+
 var (
 	targetsDiscovered = promauto.NewGaugeVec(prometheus.GaugeOpts{
 		Name: "opentelemetry_allocator_targets",
@@ -193,25 +195,24 @@ func (m *Discoverer) Reload() {
 
 // processTargetGroups processes the target groups and returns a map of targets.
 func (m *Discoverer) processTargetGroups(jobName string, groups []*targetgroup.Group, intoTargets []*Item) {
-	builder := labels.NewBuilder(labels.Labels{})
+	groupBuilder := labels.NewScratchBuilder(labelBuilderPreallocSize)
 	timer := prometheus.NewTimer(processTargetGroupsDuration.WithLabelValues(jobName))
 
 	defer timer.ObserveDuration()
 	var count float64 = 0
 	index := 0
 	for _, tg := range groups {
-		builder.Reset(labels.EmptyLabels())
+		groupBuilder.Reset()
 		for ln, lv := range tg.Labels {
-			builder.Set(string(ln), string(lv))
+			groupBuilder.Add(string(ln), string(lv))
 		}
-		groupLabels := builder.Labels()
 		for _, t := range tg.Targets {
 			count++
-			builder.Reset(groupLabels)
+			targetBuilder := groupBuilder
 			for ln, lv := range t {
-				builder.Set(string(ln), string(lv))
+				targetBuilder.Add(string(ln), string(lv))
 			}
-			item := NewItem(jobName, string(t[model.AddressLabel]), builder.Labels(), "")
+			item := NewItem(jobName, string(t[model.AddressLabel]), targetBuilder.Labels(), "")
 			intoTargets[index] = item
 			index++
 		}