ScratchBuilder is intended exactly for this use case - building labels from scratch. This results in the following performance improvement:

```
goos: linux
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator
cpu: AMD Ryzen 9 7950X3D 16-Core Processor          
                                                             │ bench_main.txt │          bench_branch2.txt           │
                                                             │     sec/op     │    sec/op     vs base                │
ProcessTargets/least-weighted/1000-32                           1069.7µ ±  6%   638.4µ ±  8%  -40.32% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                       1040.9µ ±  6%   666.9µ ± 11%  -35.92% (p=0.000 n=10)
ProcessTargets/per-node/1000-32                                 1061.3µ ± 13%   655.9µ ±  7%  -38.21% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                           7.114m ±  3%   5.741m ±  6%  -19.29% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                       7.862m ±  9%   5.665m ±  7%  -27.94% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                 7.489m ±  4%   5.599m ±  3%  -25.23% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                          72.20m ±  3%   59.38m ±  3%  -17.77% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                      72.23m ±  3%   59.07m ±  2%  -18.21% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                69.17m ±  4%   58.13m ±  4%  -15.96% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                596.9m ±  6%   563.3m ±  4%   -5.64% (p=0.001 n=10)
ProcessTargets/least-weighted/800000-32                          582.9m ±  3%   570.2m ±  6%        ~ (p=0.123 n=10)
ProcessTargets/consistent-hashing/800000-32                      585.8m ±  3%   560.9m ±  4%   -4.25% (p=0.003 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32       1.378m ±  3%   1.332m ± 13%        ~ (p=0.481 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                 1.497m ± 11%   1.297m ±  8%  -13.34% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32           1.298m ±  4%   1.291m ±  3%        ~ (p=0.796 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32      10.79m ±  1%   10.20m ±  4%   -5.46% (p=0.002 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32               10.661m ±  1%   9.928m ±  4%   -6.88% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32          10.79m ±  3%   10.08m ±  3%   -6.55% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32        110.46m ±  2%   87.10m ±  2%  -21.15% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32    112.21m ±  3%   86.00m ±  2%  -23.36% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32              110.82m ±  3%   85.83m ±  2%  -22.55% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32         804.0m ±  4%   671.0m ±  2%  -16.55% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32     787.4m ±  3%   674.4m ±  1%  -14.35% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32               781.5m ±  3%   671.8m ±  2%  -14.04% (p=0.000 n=10)
geomean                                                          28.50m         23.53m        -17.42%

                                                             │ bench_main.txt │          bench_branch2.txt           │
                                                             │      B/op      │     B/op      vs base                │
ProcessTargets/least-weighted/1000-32                            4.179Mi ± 0%   1.641Mi ± 0%  -60.74% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                        4.178Mi ± 0%   1.639Mi ± 0%  -60.77% (p=0.000 n=10)
ProcessTargets/per-node/1000-32                                  4.178Mi ± 0%   1.639Mi ± 0%  -60.77% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                           41.68Mi ± 0%   16.07Mi ± 0%  -61.43% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                       41.67Mi ± 0%   16.07Mi ± 0%  -61.44% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                 41.67Mi ± 0%   16.07Mi ± 0%  -61.43% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                          415.9Mi ± 0%   157.8Mi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                      415.9Mi ± 0%   157.8Mi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                415.9Mi ± 0%   157.8Mi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                3.249Gi ± 0%   1.233Gi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                          3.249Gi ± 0%   1.233Gi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                      3.249Gi ± 0%   1.233Gi ± 0%  -62.06% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32       4.459Mi ± 0%   2.528Mi ± 0%  -43.30% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                 4.459Mi ± 0%   2.528Mi ± 0%  -43.30% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32           4.459Mi ± 0%   2.528Mi ± 0%  -43.30% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32      44.53Mi ± 0%   25.11Mi ± 0%  -43.61% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                44.53Mi ± 0%   25.11Mi ± 0%  -43.61% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32          44.53Mi ± 0%   25.11Mi ± 0%  -43.61% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32         444.8Mi ± 0%   249.6Mi ± 0%  -43.89% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32     444.8Mi ± 0%   249.6Mi ± 0%  -43.89% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32               444.8Mi ± 0%   249.6Mi ± 0%  -43.89% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32         3.475Gi ± 0%   1.950Gi ± 0%  -43.89% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32     3.475Gi ± 0%   1.950Gi ± 0%  -43.89% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32               3.475Gi ± 0%   1.950Gi ± 0%  -43.89% (p=0.000 n=10)
geomean                                                          128.8Mi        59.91Mi       -53.48%

                                                             │ bench_main.txt │          bench_branch2.txt          │
                                                             │   allocs/op    │  allocs/op   vs base                │
ProcessTargets/least-weighted/1000-32                             3.320k ± 0%   2.107k ± 0%  -36.51% (p=0.000 n=10)
ProcessTargets/consistent-hashing/1000-32                         3.317k ± 0%   2.103k ± 0%  -36.59% (p=0.000 n=10)
ProcessTargets/per-node/1000-32                                   3.317k ± 0%   2.103k ± 0%  -36.60% (p=0.000 n=10)
ProcessTargets/least-weighted/10000-32                            32.99k ± 0%   20.64k ± 0%  -37.43% (p=0.000 n=10)
ProcessTargets/consistent-hashing/10000-32                        32.99k ± 0%   20.64k ± 0%  -37.44% (p=0.000 n=10)
ProcessTargets/per-node/10000-32                                  32.99k ± 0%   20.64k ± 0%  -37.44% (p=0.000 n=10)
ProcessTargets/least-weighted/100000-32                           328.5k ± 0%   204.5k ± 0%  -37.75% (p=0.000 n=10)
ProcessTargets/consistent-hashing/100000-32                       328.5k ± 0%   204.5k ± 0%  -37.74% (p=0.000 n=10)
ProcessTargets/per-node/100000-32                                 328.5k ± 0%   204.5k ± 0%  -37.75% (p=0.000 n=10)
ProcessTargets/per-node/800000-32                                 2.627M ± 0%   1.635M ± 0%  -37.77% (p=0.000 n=10)
ProcessTargets/least-weighted/800000-32                           2.627M ± 0%   1.635M ± 0%  -37.77% (p=0.000 n=10)
ProcessTargets/consistent-hashing/800000-32                       2.627M ± 0%   1.635M ± 0%  -37.77% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/1000-32        6.306k ± 0%   5.585k ± 0%  -11.43% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/1000-32                  6.306k ± 0%   5.585k ± 0%  -11.43% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/1000-32            6.306k ± 0%   5.585k ± 0%  -11.43% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/10000-32       62.94k ± 0%   55.53k ± 0%  -11.77% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/10000-32                 62.94k ± 0%   55.53k ± 0%  -11.78% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/10000-32           62.94k ± 0%   55.53k ± 0%  -11.78% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/100000-32          628.3k ± 0%   553.9k ± 0%  -11.85% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/100000-32      628.3k ± 0%   553.9k ± 0%  -11.85% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/100000-32                628.3k ± 0%   553.9k ± 0%  -11.86% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/least-weighted/800000-32          5.026M ± 0%   4.430M ± 0%  -11.86% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/consistent-hashing/800000-32      5.026M ± 0%   4.430M ± 0%  -11.86% (p=0.000 n=10)
ProcessTargetsWithRelabelConfig/per-node/800000-32                5.026M ± 0%   4.430M ± 0%  -11.86% (p=0.000 n=10)
geomean                                                           136.2k        101.3k       -25.66%

```